requirements.txt
streamlit
google-cloud-texttospeech
google-cloud-speech
google-generativeai
python-dotenv
pydub